requirements/base.txt
asgiref==3.5.0
attrs==21.4.0
Django==3.2.8
django-cors-headers==3.9.0
django-filter==21.1
django-guardian==2.4.0
django-taggit==2.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
inflection==0.5.1
jsonschema==4.4.0
Pillow==8.3.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1